+++ JuceLibraryCode/JuceHeader.h
@@ -47,7 +47,7 @@ namespace ProjectInfo
 {
     const char* const  projectName    = "WFS-DIY";
     const char* const  companyName    = "Pix et Bel";
-    const char* const  versionString  = "1.0.0beta50";
+    const char* const  versionString  = "1.0.0beta51";
     const int          versionNumber  = 0x10000;
 }
 #endif